backend: Changes the webcam handling to in-memory cache and sticky sessions.
Previous solution with redis works in a multi-threaded uwsgi setup. But it has some performance limitations.
This solution switches to multiple uwsgi processes with nginx-based sticky sessions and reverts back
to the in-memory cache. This should put back the raw performance of in-memory cache solution and
with sticky-sessions prevent the frame ordering mishaps.